+++ source/Messaging.Infrastructure/OutgoingMessages/CharacteristicsOfACustomerAtAnAp/CharacteristicsOfACustomerAtAnApDocumentWriter.cs
@@ -20,6 +20,7 @@ using Messaging.Application.Common;
 using Messaging.Application.OutgoingMessages;
 using Messaging.Application.OutgoingMessages.CharacteristicsOfACustomerAtAnAp;
 using Messaging.Application.OutgoingMessages.Common;
+using Messaging.Application.OutgoingMessages.Common.Xml;
 using Messaging.Infrastructure.OutgoingMessages.Common.Xml;
 
 namespace Messaging.Infrastructure.OutgoingMessages.CharacteristicsOfACustomerAtAnAp;
