+++ source/Messaging.Infrastructure/OutgoingMessages/Common/Xml/HeaderWriter.cs
@@ -16,6 +16,7 @@ using System;
 using System.Threading.Tasks;
 using System.Xml;
 using Messaging.Application.Common;
+using Messaging.Application.OutgoingMessages.Common.Xml;
 using Messaging.Domain.OutgoingMessages;
 
 namespace Messaging.Infrastructure.OutgoingMessages.Common.Xml;
